solution/requirements.txt
qutechopenql=0.8.1.dev4
quantum-decomp=1.1.1
ply==3.4